Moved group_by_attributes to calibration stages
