+++ banzai/stages.py
@@ -18,31 +18,13 @@ class Stage(abc.ABC):
     def stage_name(self):
         return '.'.join([__name__, self.__class__.__name__])
 
-    @property
-    @abc.abstractmethod
-    def group_by_attributes(self):
-        return []
-
-    def get_grouping(self, image):
-        grouping_criteria = [image.site, image.instrument, image.epoch]
-        if self.group_by_attributes:
-            grouping_criteria += [getattr(image, keyword) for keyword in self.group_by_attributes]
-        return grouping_criteria
-
-    def run_stage(self, image_set):
-        image_set = list(image_set)
-        logger.info('Running {0}'.format(self.stage_name), image=image_set[0])
-        return self.do_stage(image_set)
-
     def run(self, images):
-        images.sort(key=self.get_grouping)
-        processed_images = []
-        for _, image_set in itertools.groupby(images, self.get_grouping):
-            try:
-                processed_images += self.run_stage(image_set)
-            except Exception as e:
-                logger.error(e)
-        return processed_images
+        logger.info('Running {0}'.format(self.stage_name), image=image_set[0])
+        try:
+            images = self.do_stage(images)
+        except Exception as e:
+            logger.error(e)
+        return images
 
     @abc.abstractmethod
     def do_stage(self, images):
